+++ semantic-conventions/src/tests/semconv/model/test_error_detection.py
@@ -20,6 +20,7 @@ from opentelemetry.semconv.model.semantic_convention import (
     SemanticConventionSet,
     parse_semantic_convention_groups,
 )
+from ruamel.yaml.constructor import DuplicateKeyError
 
 
 class TestCorrectErrorDetection(unittest.TestCase):
@@ -450,7 +451,7 @@ class TestCorrectErrorDetection(unittest.TestCase):
         self.assertIn("at least one of name or prefix", msg)
         self.assertEqual(e.line, 2)
 
-    def test_condition_missing_contionally_required_attribute(self):
+    def test_condition_missing_conditionally_required_attribute(self):
         with self.assertRaises(ValidationError) as ex:
             self.open_yaml("yaml/errors/wrong_conditionally_required_no_condition.yaml")
             self.fail()
@@ -458,6 +459,11 @@ class TestCorrectErrorDetection(unittest.TestCase):
         msg = e.message.lower()
         self.assertIn("missing message for conditionally required field!", msg)
         self.assertEqual(e.line, 11)
+
+    def test_multiple_requirement_levels(self):
+        with self.assertRaises(DuplicateKeyError) as ex:
+            self.open_yaml("yaml/errors/wrong_multiple_requirement_levels.yaml")
+            self.fail()
 
     def open_yaml(self, path):
         with open(self.load_file(path), encoding="utf-8") as file: